ok bug found in start_header_vimrc Here is the bug solved:                 67 function InsertsHeader()                 68   :1                 69   g/License/s/^\(.\{1,2\}\ \)\(License:\)/\= s:printsLicence(submatch(1),submatch(2),g:path_to_licences_file)/                 70   g/^\n\(.\{1,2\}\ Purpose\)/s/^\n\(.\{1,2\}\ Purpose\)/\= submatch(1)/                 71 endfunction
We change ^\n\(. to  g/^\n\(.\{1,2\} When we  have then following
        * License :
it is whithin comment. The following it is one line comment:t
        // License :

Then trick is we take only the first character and then followed by a
space. In the previous (or secong from top) this character is doubled.
We say first character met once or twie followed by space. That's a
solution to solve the problem.