Issue met when a file already exists but has no header. iimap <c-h> <Esc>mz:call SetPosition()<cr><esc>'z
If above commznd line used It will paced header at first line of file
which is not good if it is html file for instance beacause it has
already <html> tag. It should be put after this tag.
Bug solved within  function! SetPosition()
line
:1

commented and line

:normal! dd

There is a blanck line before and, after header inserted/introduced in
the file!.